Rename InstanceDefinition to ObjectDefinition and the subnamespace to Initializer
See the discussion in #8
